US_2-01 completed and awaiting its pull request confirmation from other contributors

+++ project-app/app/controllers/SortingCentersController.scala
@@ -7,7 +7,7 @@ import com.mohiva.play.silhouette.api.{Environment, Silhouette}
 import com.mohiva.play.silhouette.impl.authenticators.CookieAuthenticator
 import forms.SortingCenterWarehouseForm
 import models.services.{SortingCenterWarehouseService, SupplyService}
-import models.{SortingCenterWarehouse, User}
+import models.{SortingCenterWarehouse, Supply, User}
 import play.api.i18n.MessagesApi
 
 import scala.concurrent.ExecutionContext.Implicits.global
@@ -56,11 +56,23 @@ class SortingCentersController @Inject()(
             idSortingCenter = UUID.randomUUID(),
             userID = request.identity.userID,
             resource = supply.resource,
-            amount = supply.amount,
+            amount = data.amount,
             inSortingCenter = false
           )
 
-        supplyService.deleteRowByID(supply.id)
+          if (supply.amount == data.amount)
+            supplyService.deleteRowByID(supply.id)
+          else {
+            val updatedSupply = Supply(
+              id = supply.id,
+              userID = supply.userID,
+              resource = supply.resource,
+              amount = supply.amount - data.amount
+            )
+
+            supplyService.save(updatedSupply)
+          }
+
           for {
             offer <- sortingCenterWarehouseService.save(offer.copy())
           } yield Redirect(routes.SortingCentersController.index())
